requirements.txt
python-dotenv>=0.20.0
requests>=2.28.1
strsim>=0.0.3
googlemaps>=4.6.0
marshmallow-dataclass>=8.5.8
arrow>=1.2.2
python-dateutil>=2.8.2
hikari>=2.0.0.dev109
hikari-lightbulb>=2.2.4
owoify-py>=2.0.0
PyYAML>=6.0
attrs>=21.4.0
uvloop>=0.16.0